feat: support wave texture
