+++ packages/vrender-core/src/render/contributions/render/path-render.ts
@@ -19,11 +19,8 @@ import type {
 } from '../../../interface';
 import { getTheme } from '../../../graphic/theme';
 import { PATH_NUMBER_TYPE } from '../../../graphic/constants';
-import { drawPathProxy, fillVisible, runFill, runStroke, strokeVisible } from './utils';
 import { PathRenderContribution } from './contributions/constants';
-import { BaseRenderContributionTime } from '../../../common/enums';
 import { BaseRender } from './base-render';
-import { mat4Allocate } from '../../../allocator/matrix-allocate';
 import {
   defaultPathBackgroundRenderContribution,
   defaultPathTextureRenderContribution
