+++ packages/vrender-core/src/render/contributions/render/utils.ts
@@ -70,86 +70,6 @@ export function rectStrokeVisible(opacity: number, strokeOpacity: number, width:
   return opacity * strokeOpacity > 0 && width !== 0 && height !== 0;
 }
 
-export function drawPathProxy(
-  graphic: IGraphic,
-  context: IContext2d,
-  x: number,
-  y: number,
-  drawContext: IDrawContext,
-  params?: IGraphicRenderDrawParams,
-  fillCb?: (
-    ctx: IContext2d,
-    markAttribute: Partial<IMarkAttribute & IGraphicAttribute>,
-    themeAttribute: IThemeAttribute
-  ) => boolean,
-  strokeCb?: (
-    ctx: IContext2d,
-    markAttribute: Partial<IMarkAttribute & IGraphicAttribute>,
-    themeAttribute: IThemeAttribute
-  ) => boolean
-) {
-  if (!graphic.pathProxy) {
-    return false;
-  }
-
-  const themeAttributes = (getTheme(graphic, params?.theme) as any)[graphic.type.replace('3d', '')];
-
-  const {
-    fill = themeAttributes.fill,
-    stroke = themeAttributes.stroke,
-    opacity = themeAttributes.opacity,
-    fillOpacity = themeAttributes.fillOpacity,
-    lineWidth = themeAttributes.lineWidth,
-    strokeOpacity = themeAttributes.strokeOpacity,
-    visible = themeAttributes.visible,
-    x: originX = themeAttributes.x,
-    y: originY = themeAttributes.y
-  } = graphic.attribute;
-  // 不绘制或者透明
-  const fVisible = fillVisible(opacity, fillOpacity, fill);
-  const sVisible = strokeVisible(opacity, strokeOpacity);
-  const doFill = runFill(fill);
-  const doStroke = runStroke(stroke, lineWidth);
-
-  if (!visible) {
-    return true;
-  }
-
-  if (!(doFill || doStroke)) {
-    return true;
-  }
-
-  // 如果存在fillCb和strokeCb，那就不直接跳过
-  if (!(fVisible || sVisible || fillCb || strokeCb)) {
-    return true;
-  }
-
-  context.beginPath();
-  const path = typeof graphic.pathProxy === 'function' ? graphic.pathProxy(graphic.attribute) : graphic.pathProxy;
-  renderCommandList(path.commandList, context, x, y);
-
-  // shadow
-  context.setShadowBlendStyle && context.setShadowBlendStyle(graphic, graphic.attribute, themeAttributes);
-
-  if (doStroke) {
-    if (strokeCb) {
-      strokeCb(context, graphic.attribute, themeAttributes);
-    } else if (sVisible) {
-      context.setStrokeStyle(graphic, graphic.attribute, x - originX, y - originY, themeAttributes);
-      context.stroke();
-    }
-  }
-  if (doFill) {
-    if (fillCb) {
-      fillCb(context, graphic.attribute, themeAttributes);
-    } else if (fVisible) {
-      context.setCommonStyle(graphic, graphic.attribute, x - originX, y - originY, themeAttributes);
-      context.fill();
-    }
-  }
-  return true;
-}
-
 /**
  * 部分源码参考 https://github.com/d3/d3-shape/
  * Copyright 2010-2022 Mike Bostock
